+++ src/main/java/lagatrix/client/tools/DrawFactory.java
@@ -2,8 +2,10 @@ package lagatrix.client.tools;
 
 import java.awt.Graphics;
 import java.awt.Graphics2D;
+import java.awt.Rectangle;
 import java.awt.RenderingHints;
 import java.awt.Shape;
+import java.awt.geom.Area;
 import java.awt.geom.RoundRectangle2D;
 import javax.swing.JComponent;
 
@@ -38,17 +40,34 @@ public class DrawFactory {
                 RenderingHints.VALUE_ANTIALIAS_ON);
     }
     
-    /**
+     /**
      * This method performs corner bending to the components.
-     * 
+     *
      * @param g The grafic object.
      * @param radius Degree of curvature.
      */
     public void roundComponent(Graphics g, int radius) {
         Graphics2D g2 = (Graphics2D) g;
-        Shape shape = new RoundRectangle2D.Float(0, 0, component.getWidth() 
-                - 1, component.getHeight() - 1, radius, radius);
-        
+        Shape shape = new RoundRectangle2D.Float(0, 0,
+                component.getWidth(), component.getHeight(), radius, radius);
+       
         g2.setClip(shape);
+    }
+   
+    /**
+     * This method performs corner in right side bending to the components.
+     *
+     * @param g The grafic object.
+     * @param radius Degree of curvature.
+     */
+    public void roundRight(Graphics g, int radius) {
+        Graphics2D g2 = (Graphics2D) g;
+        Area shape1 = new Area(new RoundRectangle2D.Float(0, 0,
+                component.getWidth(), component.getHeight(), radius, radius));
+        Area shape2 = new Area(new Rectangle(100, 0, component.getWidth() - 100 , component.getHeight()));
+       
+        shape1.add(shape2);
+       
+        g2.setClip(shape1);
     }
 }